+++ release_notes.md
@@ -9,3 +9,5 @@
     * Add some convenient POJOs. (#774)
     * Support multiple Azure SignalR Service instances. (#566)
     * Fix a bug that group name is required when group action is "RemoveAll". (#347)
+- Bug fix: Handle = character in cookie value (#838)
+- Error handling for in-proc SDK usage inside isolated app. (#824)